+++ linkerd/app/outbound/src/test_util.rs
@@ -60,7 +60,7 @@ pub(crate) fn runtime() -> (ProxyRuntime, drain::Signal) {
     let (tap, _) = tap::new();
     let (metrics, _) = metrics::Metrics::new(std::time::Duration::from_secs(10));
     let runtime = ProxyRuntime {
-        identity: linkerd_meshtls_rustls::creds::default_for_test().1,
+        identity: linkerd_meshtls::creds::default_for_test().1,
         metrics: metrics.proxy,
         tap,
         span_sink: None,
