--- linkerd/meshtls/rustls/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "linkerd-meshtls-rustls"
-version = { workspace = true }
-authors = { workspace = true }
-license = { workspace = true }
-edition = "2018"
-publish = { workspace = true }
-
-[features]
-aws-lc-fips = ["tokio-rustls/fips"]
-test-util = ["linkerd-tls-test-util"]
-
-[dependencies]
-futures = { version = "0.3", default-features = false }
-rustls-pemfile = "2.2"
-rustls-webpki = { version = "0.103.4", default-features = false, features = ["std", "aws-lc-rs"] }
-thiserror = "2"
-tokio = { version = "1", features = ["macros", "rt", "sync"] }
-tokio-rustls = { workspace = true, features = ["aws-lc-rs"] }
-tracing = { workspace = true }
-
-linkerd-dns-name = { path = "../../dns/name" }
-linkerd-error = { path = "../../error" }
-linkerd-io = { path = "../../io" }
-linkerd-identity = { path = "../../identity" }
-linkerd-stack = { path = "../../stack" }
-linkerd-tls = { path = "../../tls" }
-linkerd-tls-test-util = { path = "../../tls/test-util", optional = true }
-linkerd-meshtls-verifier = { path = "../verifier" }
-
-[dev-dependencies]
-linkerd-tls-test-util = { path = "../../tls/test-util" }
