--- linkerd/meshtls/rustls/src/lib.rs
@@ -1,14 +1,0 @@
-#![deny(rust_2018_idioms, clippy::disallowed_methods, clippy::disallowed_types)]
-#![forbid(unsafe_code)]
-
-mod backend;
-mod client;
-pub mod creds;
-mod server;
-#[cfg(test)]
-mod tests;
-
-pub use self::{
-    client::{ClientIo, Connect, ConnectFuture, NewClient},
-    server::{Server, ServerIo, TerminateFuture},
-};
