+++ linkerd/meshtls/src/backend/aws_lc.rs
@@ -9,25 +9,25 @@ pub fn default_provider() -> CryptoProvider {
     provider.cipher_suites = TLS_SUPPORTED_CIPHERSUITES.to_vec();
     provider.kx_groups = kx_groups();
     provider.signature_verification_algorithms = *SUPPORTED_SIG_ALGS;
-    #[cfg(feature = "aws-lc-fips")]
+    #[cfg(feature = "rustls-aws-lc-fips")]
     assert!(provider.fips());
     provider
 }
 
-#[cfg(not(feature = "aws-lc-fips"))]
+#[cfg(not(feature = "rustls-aws-lc-fips"))]
 pub static TLS_SUPPORTED_CIPHERSUITES: &[rustls::SupportedCipherSuite] = &[
     aws_lc_rs::cipher_suite::TLS13_AES_128_GCM_SHA256,
     aws_lc_rs::cipher_suite::TLS13_AES_256_GCM_SHA384,
     aws_lc_rs::cipher_suite::TLS13_CHACHA20_POLY1305_SHA256,
 ];
 // Prefer aes-256-gcm if fips is enabled
-#[cfg(feature = "aws-lc-fips")]
+#[cfg(feature = "rustls-aws-lc-fips")]
 static TLS_SUPPORTED_CIPHERSUITES: &[rustls::SupportedCipherSuite] = &[
     aws_lc_rs::cipher_suite::TLS13_AES_256_GCM_SHA384,
     aws_lc_rs::cipher_suite::TLS13_AES_128_GCM_SHA256,
 ];
 
-#[cfg(not(feature = "aws-lc-fips"))]
+#[cfg(not(feature = "rustls-aws-lc-fips"))]
 fn kx_groups() -> Vec<&'static dyn SupportedKxGroup> {
     vec![
         aws_lc_rs::kx_group::X25519MLKEM768,
@@ -36,7 +36,7 @@ fn kx_groups() -> Vec<&'static dyn SupportedKxGroup> {
         aws_lc_rs::kx_group::SECP384R1,
     ]
 }
-#[cfg(feature = "aws-lc-fips")]
+#[cfg(feature = "rustls-aws-lc-fips")]
 fn kx_groups() -> Vec<&'static dyn SupportedKxGroup> {
     vec![
         aws_lc_rs::kx_group::SECP256R1MLKEM768,
